Data.Foldable1.WithIndex

Description

Indexed non-empty Foldables.

Indexed Foldables

class (Foldable1 f, FoldableWithIndex i f) => Foldable1WithIndex i f | f -> i where

A non-empty container that supports folding with an additional index.

Minimal complete definition

ifoldMap1 | ifoldrMap1

Methods

ifoldMap1 :: Semigroup m => (i -> a -> m) -> f a -> m

Map each element of the structure to a semigroup, and combine the results.

ifoldMap1' :: Semigroup m => (i -> a -> m) -> f a -> m

A variant of ifoldMap1 that is strict in the accumulator.

ifoldrMap1 :: (i -> a -> b) -> (i -> a -> b -> b) -> f a -> b

Generalized ifoldr1.

ifoldlMap1' :: (i -> a -> b) -> (i -> b -> a -> b) -> f a -> b

Generalized ifoldl1'.

ifoldlMap1 :: (i -> a -> b) -> (i -> b -> a -> b) -> f a -> b

Generalized ifoldl1.

ifoldrMap1' :: (i -> a -> b) -> (i -> a -> b -> b) -> f a -> b

Generalized ifoldr1'.